Validate methods before creating compiled method nodes
Copilot-Session: de33062d-4d14-4f93-a1bb-3d6036a7cabd
